Crate nsring

Nifty Simple Fixed Circular Buffer (nsring) provides macros to rapidly create circular buffer data structure on the stack.

It also provides a no tail buffer (Manx buffer) for rapid data accumulation for history and logging.

Reference(s)

Manx buffer

Usage

#[macro_use] extern crate nsring;
 
nsring::ring!(UsizeRB, usize, 10); // Use macro to circular buffer structure.
 
fn main() {
    let mut rb = UsizeRB::new();
    rb.push(5);
    assert_eq!(*rb.pop().unwrap(), 5);
}

Macros

fcb_generic

Create a generic circular buffer reusable with multiple types.

manx

Create a manx circular buffer for a single type.

manx_generic

Create a generic manx circular buffer reusable with multiple types.

ring

Create a fixed ring buffer for specific type.